// src/app/models/project/policy/procost.ts
export class ProcostModel {
    constructor() {
             
    }
    
    procost_id: string = "1";
    procost_code: string = "";
    procost_name_th: string = "";
    procost_name_en: string = "";
    procost_type: string = "";
    procost_auto: boolean = false;
    procost_itemcode: string = "";
    company_code: string = "";
    modified_by: string = "";
    modified_date: string = "";
    index: number = 0;
    select: boolean = false;    
  }
  