// src/app/models/project/project_withdraw.ts
export class ProjectwithdrawModel {
    constructor() {
     
    }
    
    
    prowithdraw_id: number = 0;
    prowithdraw_date: Date = new Date();

    project_code: string = "";
    projob_code: string = "";
    worker_code: string = "";
   

    modified_by: string = "";
    modified_date: string = "";
    index: number = 0;

    //-- display
    allow1: number = 0;
    allow2: number = 0;
    allow3: number = 0;
    allow4: number = 0;
    allow5: number = 0;
    allow6: number = 0;
    allow7: number = 0;
    allow8: number = 0;
    allow9: number = 0;
    allow10: number = 0;
    allow_emp: number = 0;
    allow_total: number = 0;

    worker_name_th: string = "";
    worker_name_en: string = "";

    project_name_th: string = "";
    project_name_en: string = "";

   
  }
  